Added the ability to persist top organisms to disk
I'm not adding a disk-based store, per-se, but I want to be able to
persist the best n organisms after a test run and then have those
pre-populate the store before the cthonic initialization of the store.
Eventually, I will use the chromosomes from these organisms in my
ChromosomeFactory.

This will lead to several things that have nothing to do with disk
persistence, specifically around writing organisms based on technical
analysis patterns, and using those to pre-populate a store.